refactor the js action structure and include a minified compiled js for action use
