+++ .github/actions/cache_cargo/.gitignore
@@ -1,0 +1,2 @@
+src/node_modules
+**/node_modules
