+++ .github/actions/cache_cargo/src/README.md
@@ -1,0 +1,7 @@
+```
+npm i -g @vercel/ncc
+```
+
+```
+ncc build cache_dirs.js --minify --source-map --out ../dist
+```
